more funcs moved
